@@ -5,31 +5,26 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "ghpage": "npm run build:page && cp ./docs/index.html ./docs/404.html",
-    "build:page": "ng build --output-path docs --base-href /angular-maze/",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "watch": "ng build --watch --configuration development",
+    "build:gh-pages": "ng build --base-href /angular-maze/"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.1.3",
-    "@angular/cdk": "^15.1.3",
-    "@angular/common": "^15.1.3",
-    "@angular/compiler": "^15.1.3",
-    "@angular/core": "^15.1.3",
-    "@angular/forms": "^15.1.3",
-    "@angular/platform-browser": "^15.1.3",
-    "@angular/platform-browser-dynamic": "^15.1.3",
-    "@angular/router": "^15.1.3",
+    "@angular/common": "^19.2.0",
+    "@angular/compiler": "^19.2.0",
+    "@angular/core": "^19.2.0",
+    "@angular/forms": "^19.2.0",
+    "@angular/platform-browser": "^19.2.0",
+    "@angular/platform-browser-dynamic": "^19.2.0",
+    "@angular/router": "^19.2.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.12.0"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.4",
-    "@angular/cli": "^15.1.4",
-    "@angular/compiler-cli": "^15.1.3",
-    "typescript": "~4.9.5"
+    "@angular-devkit/build-angular": "^19.2.5",
+    "@angular/cli": "^19.2.5",
+    "@angular/compiler-cli": "^19.2.0",
+    "typescript": "~5.7.2"
   }
 }
